+++ kamehouse-shell/bin/csv/csv-kamehouse-shell.sh
@@ -5,6 +5,8 @@
 source ${HOME}/programs/kamehouse-shell/bin/functions/common-functions.sh
 if [ "$?" != "0" ]; then echo "Error importing common-functions.sh" ; exit 99 ; fi
 
+importKamehouse functions/kamehouse/get-kamehouse-shell-path-content-functions.sh
+
 initKameHouseShellEnv() {
   LOG=DISABLED
 }
@@ -14,29 +16,7 @@ initScriptEnv() {
 }
 
 mainProcess() {  
-  # List all files
-  SCRIPTS_PATH=$(find ${BASE_DIR} -name '.*' -prune -o -type f)
-
-  # Filter bashrc
-  SCRIPTS_PATH=$(echo "$SCRIPTS_PATH" | grep -v /bashrc) 
-  # Filter docker container scripts
-  SCRIPTS_PATH=$(echo "$SCRIPTS_PATH" | grep -v /docker/docker-container)
-  SCRIPTS_PATH=$(echo "$SCRIPTS_PATH" | grep -v /docker/release/java8-release/bin)
-  SCRIPTS_PATH=$(echo "$SCRIPTS_PATH" | grep -v /docker/release/java8-release/docker)
-  SCRIPTS_PATH=$(echo "$SCRIPTS_PATH" | grep -v /docker/release/java11-release/bin)
-  SCRIPTS_PATH=$(echo "$SCRIPTS_PATH" | grep -v /docker/release/java11-release/docker)    
-  
-  # Filter -functions.sh scripts
-  SCRIPTS_PATH=$(echo "$SCRIPTS_PATH" | grep -v -e '\-functions\.sh$') 
-
-  # Keep only .sh scripts
-  SCRIPTS_PATH=$(echo "$SCRIPTS_PATH" | grep -e '\.sh$') 
-
-  # Replace \n with :  
-  SCRIPTS_PATH=$(echo "$SCRIPTS_PATH" | tr '\n' ':')
-
-  # Remove last :
-  SCRIPTS_PATH=$(echo "$SCRIPTS_PATH" | sed '$s/.$//')
+  local SCRIPTS_PATH=`getKameHouseShellPathContent "${BASE_DIR}" "f"`
 
   # Convert : to ,
   SCRIPTS_PATH=$(echo "$SCRIPTS_PATH" | tr ':' ',')