+++ kamehouse-shell/bin/functions/urlencode/urlencode-functions.sh
@@ -1,4 +1,3 @@
-# Custom non-native bash functions
 
 urlencode() {
     # urlencode <string>